refactor : routine component
