+++ frontend/pages/routine/routineIng.jsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import * as RS from "@/components/_styled/routineStyled";
-import Timer from '@/components/routine/Timer';
-import TimerInfo from '@/components/routine/TimerInfo';
+import Timer from '@/components/routine/routineIng/Timer';
+import TimerInfo from '@/components/routine/routineIng/TimerInfo';
 export default function routineIng() {
 
     return (
